+++ ydb/public/sdk/cpp/client/ydb_debug/client.h
@@ -2,7 +2,7 @@
 
 #include <ydb/public/sdk/cpp/client/ydb_driver/driver.h>
 
-namespace NYdb::inline V2::NDebug {
+namespace NYdb::inline V3::NDebug {
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -41,6 +41,13 @@ public:
     {}
 };
 
+class TActorChainPingResult: public TStatus {
+public:
+    TActorChainPingResult(TStatus&& status)
+        : TStatus(std::move(status))
+    {}
+};
+
 ////////////////////////////////////////////////////////////////////////////////
 
 using TAsyncPlainGrpcPingResult = NThreading::TFuture<TPlainGrpcPingResult>;
@@ -48,6 +55,7 @@ using TAsyncGrpcProxyPingResult = NThreading::TFuture<TGrpcProxyPingResult>;
 using TAsyncKqpProxyPingResult = NThreading::TFuture<TKqpProxyPingResult>;
 using TAsyncSchemeCachePingResult = NThreading::TFuture<TSchemeCachePingResult>;
 using TAsyncTxProxyPingResult = NThreading::TFuture<TTxProxyPingResult>;
+using TAsyncActorChainPingResult = NThreading::TFuture<TActorChainPingResult>;
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -56,6 +64,12 @@ struct TGrpcProxyPingSettings : public TOperationRequestSettings<TGrpcProxyPingS
 struct TKqpProxyPingSettings : public TOperationRequestSettings<TKqpProxyPingSettings> {};
 struct TSchemeCachePingSettings : public TOperationRequestSettings<TSchemeCachePingSettings> {};
 struct TTxProxyPingSettings : public TOperationRequestSettings<TTxProxyPingSettings> {};
+
+struct TActorChainPingSettings : public TOperationRequestSettings<TActorChainPingSettings> {
+    FLUENT_SETTING_DEFAULT(size_t, ChainLength, 10);
+    FLUENT_SETTING_DEFAULT(size_t, WorkUsec, 5);
+    FLUENT_SETTING_DEFAULT(bool, NoTailChain, false);
+};
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -75,9 +89,11 @@ public:
     TAsyncSchemeCachePingResult PingSchemeCache(const TSchemeCachePingSettings& settings);
     TAsyncTxProxyPingResult PingTxProxy(const TTxProxyPingSettings& settings);
 
+    TAsyncActorChainPingResult PingActorChain(const TActorChainPingSettings& settings);
+
 private:
     class TImpl;
     std::shared_ptr<TImpl> Impl_;
 };
 
-} // namespace NYdb::NDebug
+} // namespace NYdb::V3::NDebug